requirements.txt
langchain==0.3.27
langchain_community==0.3.29
dotenv==0.9.9
beautifulsoup4==4.13.5
html2text==2025.4.15
pypdf==6.1.0
langchain-unstructured==0.1.6
unstructured==0.18.15
langchain-text-splitters==0.3.11
weaviate-client==4.16.9
langchain_huggingface==0.3.1
langchain-weaviate==0.0.5
langchain-groq==0.3.8
sentence-transformers==5.1.1
lark==1.3.0
langchain_openai==0.3.33
langsmith==0.4.31
langgraph==0.6.8
